server/pyproject.toml
[project]
name = "pulse-server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "python-socketio>=5.11.3",
    "uvicorn[standard]>=0.30.1",
    "pydantic>=2.7.3",
    "aiofiles>=25.1.0",
]

